Refactor: api분리 #2
