+++ package.json
@@ -11,7 +11,6 @@
   "dependencies": {
     "@next/bundle-analyzer": "^13.4.10",
     "@tanstack/react-query": "^4.32.1",
-    "@tanstack/react-query-next-experimental": "^5.0.0-alpha.80",
     "@types/node": "20.4.2",
     "@types/react": "18.2.15",
     "@types/react-dom": "18.2.7",
